requirements.txt
aiocron==1.8
asyncio==3.4.3
croniter==3.0.3
pyaes==1.6.1
pyasn1==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
rsa==4.9
six==1.16.0
Telethon==1.36.0
tzlocal==5.2
